backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
google-genai>=1.16.0
python-dotenv
aiofiles
pillow
